pyproject.toml
[project]
name = "new-aprenda-ai-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Luiz-Trindade",email = "luiz.gabriel.m.trindade@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "uvicorn[all] (>=0.34.2,<0.35.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
